+++ src/main/java/org/teleight/teleightbots/bot/manager/BotManagerImpl.java
@@ -8,8 +8,8 @@ import org.teleight.teleightbots.bot.TelegramBot;
 import org.teleight.teleightbots.bot.WebhookTelegramBot;
 import org.teleight.teleightbots.bot.settings.LongPollingBotSettings;
 import org.teleight.teleightbots.bot.settings.WebhookBotSettings;
-import org.teleight.teleightbots.updateprocessor.webhook.WebhookServer;
-import org.teleight.teleightbots.updateprocessor.webhook.WebhookServerConfig;
+import org.teleight.teleightbots.webhook.WebhookServer;
+import org.teleight.teleightbots.webhook.WebhookServerConfig;
 
 import java.io.IOException;
 import java.util.Collection;
